% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexminer.R
\name{clean_json}
\alias{clean_json}
\title{Clean raw imported item}
\usage{
clean_json(json)
}
\arguments{
\item{json}{raw json item of api request - generated with import_item function}
}
\value{
clean tibble of item with value, date, server, faction, item ID and item name
}
\description{
Clean raw imported item
}
\examples{

bl_raw <- import_item("black lotus", "patchwerk", "horde")
clean_json(bl_raw)

}
