% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexminer.R
\name{import_item}
\alias{import_item}
\title{Import Item}
\usage{
import_item(item_string, server, faction)
}
\arguments{
\item{item_string}{any correct written string of an item}

\item{server}{string of server, eg Patchwerk}

\item{faction}{string of faction, horde or alliance}
}
\value{
raw json of api request
}
\description{
This function is a basic import of the raw json from the nexhyub API
}
\examples{

import_item("black lotus", "patchwerk", "horde")

}
