% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexminer.R
\name{smooth_item}
\alias{smooth_item}
\title{creates smooth plot for price development}
\usage{
smooth_item(item_clean)
}
\arguments{
\item{item_clean}{raw json item of api request - generated with import_item function}
}
\value{
plot of item value including smooth curve
}
\description{
creates smooth plot for price development
}
\examples{
bl_raw <- import_item("black lotus", server, fraction)
bl_clean <- clean_json(bl_raw)
smooth_item(bl_clean)
}
